requirements.txt
Django==4.0.0
cx-Oracle==8.3.0
requests==2.30.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
django-environ==0.10.0
django-filter==22.1
